Change data member representing a file descriptor from an unsigned short to an int.  This is so that we can set invalid file descriptors to -1.

+++ src/condor_ckpt/file_state.h
@@ -6,7 +6,7 @@
 
 typedef unsigned int	Bit;
 typedef int				BOOL;
-typedef unsigned short	FD;
+typedef int				FD;
 
 class OpenFileTable;
